refactor(rolldown_fs): export functionalities using `features`
