+++ crates/rolldown_binding_wasm/Cargo.toml
@@ -14,7 +14,7 @@ crate-type = ["cdylib"]
 [dependencies]
 wasm-bindgen = "0.2.74"
 rolldown     = { path = "../rolldown/" }
-rolldown_fs  = { path = "../rolldown_fs" }
+rolldown_fs  = { path = "../rolldown_fs", features = ["memory"], default-features = false }
 tokio        = { workspace = true, default-features = false, features = ["rt", "macros", "sync"] }
 
 sugar_path               = { workspace = true }
