+++ crates/rolldown_fs/Cargo.toml
@@ -8,8 +8,13 @@ repository.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+# Defines a feature named `webp` that does not enable any other features.
+default = ["os"]
+os      = []
+memory  = []
+
 [dependencies]
-anyhow       = { workspace = true }
 vfs          = "0.10.0"
 oxc_resolver = { workspace = true }
 dunce        = "1.0.4"              # Normalize Windows paths to the most compatible format, avoiding UNC where possible
